arange sum

+++ JAxtar/hash.py
@@ -56,8 +56,9 @@ def dataclass_hashing(x, seed):
     x is a dataclass
     """
     def _h(x):
+        # sum of hash * index for collision
         hashs = hashing(x, seed)
-        return jnp.sum(hashs * jnp.arange(1, len(hashs) + 1)) # sum of hash * index for collision
+        return jnp.sum(hashs * jnp.arange(1, len(hashs) + 1), dtype=jnp.uint32)
     tree_hash = jax.tree_map(_h, x)
     flattened_sum_hash = sum(jax.tree_leaves(tree_hash))
     return flattened_sum_hash
